fix: Open waiting account details on tap
A tap on the waiting row did not open the approval tooltip. The row
is now a button. Setup files stay under the ESLint line budget.
