+++ web_src/src/pages/factories/pages/onboarding/FirstRunSetup.spec.tsx
@@ -1,124 +1,20 @@
-import { render, renderHook, screen, waitFor } from "@testing-library/react";
+import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import type * as ReactRouterDom from "react-router";
-import { MemoryRouter } from "react-router";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-
-import type { FactoriesFactory } from "@/api-client";
+import { beforeEach, describe, expect, it } from "vitest";
 
 import { FIRST_RUN_COPY } from "./first-run/firstRunCopy";
-import { FirstRunSetup } from "./FirstRunSetup";
-import { useOnboardingSetupState, type OnboardingSetupApi } from "./useOnboardingSetupState";
-import type { useOnboardingPageModel } from "./useOnboardingPageModel";
-
-type OnboardingPageModel = ReturnType<typeof useOnboardingPageModel>;
-
-let factory: FactoriesFactory;
-let factories: FactoriesFactory[];
-
-vi.mock("../../layout/factoriesLayoutContext", () => ({
-  useFactoriesLayout: () => ({
-    organizationId: "org-1",
-    factoryId: "factory-1",
-    factoryKey: "PAY",
-    factory,
-    factories,
-  }),
-}));
-
-vi.mock("@/contexts/useAccount", () => ({
-  useAccount: () => ({ account: { id: "account-1", name: "Ada Lovelace", email: "ada@example.com" } }),
-}));
-
-vi.mock("@/hooks/useMe", () => ({
-  useMe: () => ({ data: { id: "user-1" } }),
-}));
-
-vi.mock("@/posthog", () => ({ posthog: { reset: vi.fn() } }));
-
-// The install-request recheck and the in-place bind need a query client and
-// the network; the flow tests cover the screens only.
-vi.mock("@/hooks/useRecheckGitHubInstallRequest", () => ({
-  useRecheckGitHubInstallRequest: vi.fn(),
-}));
-
-const bindMutate = vi.fn();
-
-vi.mock("@/hooks/useBindGitHubInstallation", () => ({
-  useBindGitHubInstallation: () => ({ mutate: bindMutate, isPending: false, variables: undefined }),
-}));
-
-const navigateSpy = vi.fn();
-
-let accountOrganizations: Array<{ id: string; name: string; slug?: string }>;
-
-vi.mock("@/hooks/useAccountOrganizations", () => ({
-  useAccountOrganizations: () => ({ data: accountOrganizations, refetch: vi.fn() }),
-}));
-
-vi.mock("react-router", async () => {
-  const actual = await vi.importActual<typeof ReactRouterDom>("react-router");
-  return {
-    ...actual,
-    useNavigate: () => navigateSpy,
-  };
-});
-
-// The agent step reports organization spend, which this flow test does not use.
-vi.mock("./AgentStep", () => ({
-  AgentStep: () => <div data-testid="agent-step" />,
-}));
-
-function setupState(): OnboardingSetupApi {
-  const { result } = renderHook(() => useOnboardingSetupState("Payments Service", { simulateDiscovery: false }));
-  return result.current;
-}
-
-function pageModel(overrides: Partial<OnboardingPageModel> = {}): OnboardingPageModel {
-  return {
-    setup: setupState(),
-    hostedAgentReady: false,
-    openSection: "issues",
-    setOpenSection: vi.fn(),
-    requestConnect: vi.fn(),
-    refreshGithubConnections: vi.fn().mockResolvedValue(undefined),
-    githubConnectionsLoading: false,
-    requestPrivateGitHubConnect: vi.fn(),
-    offersPrivateGitHubAppSetup: false,
-    createVcsConnection: vi.fn(),
-    selectVcsConnection: vi.fn().mockResolvedValue(true),
-    githubConnections: { name: "github", allInstances: [], readyInstances: [] },
-    selectedVcsConnectionId: "github-1",
-    requestConfigure: vi.fn(),
-    integrationDialogs: <></>,
-    repositories: ["acme/payments-service"],
-    repositoriesLoading: false,
-    repositoriesError: null,
-    canConfigureWorkspace: true,
-    saving: false,
-    saveName: vi.fn().mockResolvedValue(true),
-    saveRepository: vi.fn().mockResolvedValue(true),
-    saveIssues: vi.fn().mockResolvedValue(true),
-    finish: vi.fn(),
-    ...overrides,
-  };
-}
-
-function renderSetup(model: OnboardingPageModel, path = "/org-1/workspaces/PAY/setup?step=issues") {
-  render(
-    <MemoryRouter initialEntries={[path]}>
-      <FirstRunSetup model={model} />
-    </MemoryRouter>,
-  );
-}
+import {
+  navigateSpy,
+  pageModel,
+  renderSetup,
+  resetSetupFixtures,
+  setupFixtures,
+  setupState,
+} from "./FirstRunSetup.testHelpers";
 
 describe("FirstRunSetup", () => {
   beforeEach(() => {
-    factory = { id: "factory-1", onboarding: { vcsIntegrationId: "github-1" } };
-    factories = [factory];
-    accountOrganizations = [{ id: "org-1", name: "Acme" }];
-    navigateSpy.mockClear();
-    bindMutate.mockReset();
+    resetSetupFixtures();
   });
 
   it("finishes setup from the ticket screen when hosted credentials cover the agent", async () => {
@@ -155,7 +51,7 @@ describe("FirstRunSetup", () => {
   });
 
   it("asks to connect GitHub when the workspace has not saved a connection", () => {
-    factory = { id: "factory-1", onboarding: {} };
+    setupFixtures.factory = { id: "factory-1", onboarding: {} };
 
     renderSetup(
       pageModel({
@@ -218,109 +114,6 @@ describe("FirstRunSetup", () => {
     expect(screen.queryByTestId("first-run-connect")).not.toBeInTheDocument();
   });
 
-  /** The workspace's bound connection, with the picker data a bind keeps. */
-  function boundConnectionModel(selectVcsConnection: OnboardingPageModel["selectVcsConnection"]) {
-    const boundInstance = {
-      metadata: { id: "github-1", name: "github-acme", integrationName: "github" },
-      status: {
-        state: "ready",
-        metadata: {
-          owner: "acme",
-          startedByUserID: "user-1",
-          startedByGitHubLogin: "forestileao",
-          state: "csrf",
-          githubApp: { slug: "superplane" },
-          pendingInstallations: [
-            { id: "11", accountLogin: "acme" },
-            { id: "22", accountLogin: "octo" },
-          ],
-        },
-      },
-    };
-    return pageModel({
-      openSection: "vcs",
-      setup: { ...setupState(), vcsReady: true },
-      selectedVcsConnectionId: "github-1",
-      selectVcsConnection,
-      githubConnections: {
-        name: "github",
-        readyInstances: [boundInstance],
-        allInstances: [boundInstance],
-      },
-    });
-  }
-
-  it("reopens the account picker for the workspace's bound connection", () => {
-    renderSetup(boundConnectionModel(vi.fn().mockResolvedValue(true)), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    expect(screen.getByTestId("first-run-github-account-picker")).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("octo") })).toBeInTheDocument();
-    expect(screen.getByTestId("first-run-github-signed-in-as")).toHaveTextContent(
-      FIRST_RUN_COPY.connect.signedInAs("forestileao"),
-    );
-    expect(screen.queryByTestId("first-run-github-connected")).not.toBeInTheDocument();
-  });
-
-  // Get started always opens the Connect GitHub page, even when picker data
-  // exists. The user picks the GitHub account on every forward pass, because
-  // many people stay signed in to two GitHub accounts.
-  it("opens the Connect GitHub page from Get started even when picker data exists", async () => {
-    const user = userEvent.setup();
-
-    renderSetup(boundConnectionModel(vi.fn().mockResolvedValue(true)), "/org-1/workspaces/PAY/setup");
-
-    expect(screen.getByTestId("first-run-welcome")).toBeInTheDocument();
-    await user.click(screen.getByTestId("first-run-get-started"));
-
-    expect(screen.getByTestId("first-run-connect-github")).toBeInTheDocument();
-    expect(screen.queryByTestId("first-run-github-account-picker")).not.toBeInTheDocument();
-  });
-
-  // Back walks the exact pages in reverse order: account picker, Connect
-  // GitHub page, welcome.
-  it("walks back from the picker to the Connect GitHub page and then to welcome", async () => {
-    const user = userEvent.setup();
-
-    renderSetup(boundConnectionModel(vi.fn().mockResolvedValue(true)), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    expect(screen.getByTestId("first-run-github-account-picker")).toBeInTheDocument();
-    await user.click(screen.getByTestId("first-run-back"));
-
-    expect(screen.getByTestId("first-run-connect-github")).toBeInTheDocument();
-    expect(screen.queryByTestId("first-run-github-account-picker")).not.toBeInTheDocument();
-    await user.click(screen.getByTestId("first-run-back"));
-
-    expect(screen.getByTestId("first-run-welcome")).toBeInTheDocument();
-  });
-
-  it("reopens the account picker when the user goes back from the repository screen", async () => {
-    const user = userEvent.setup();
-    const selectVcsConnection = vi.fn().mockResolvedValue(true);
-    bindMutate.mockImplementation((_vars: unknown, options: { onSuccess?: () => void }) => {
-      options.onSuccess?.();
-    });
-
-    renderSetup(boundConnectionModel(selectVcsConnection), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    await user.click(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("octo") }));
-    expect(await screen.findByTestId("first-run-choose")).toBeInTheDocument();
-
-    await user.click(screen.getByTestId("first-run-back"));
-    expect(screen.getByTestId("first-run-github-account-picker")).toBeInTheDocument();
-  });
-
-  // A GitHub round trip reloads the page, so the picker data arrives after
-  // the first render. The screen must not flash the connect button first.
-  it("shows a placeholder on the connect screen while the connection list loads", () => {
-    renderSetup(
-      pageModel({ openSection: "vcs", githubConnectionsLoading: true }),
-      "/org-1/workspaces/PAY/setup?step=vcs",
-    );
-
-    expect(screen.getByTestId("first-run-connect-loading")).toBeInTheDocument();
-    expect(screen.queryByTestId("first-run-connect-github")).not.toBeInTheDocument();
-  });
-
   // A refresh of the repository list must not show cached repositories from
   // an earlier connection while the fresh list loads.
   it("shows a placeholder on the repository screen while the list refreshes", () => {
@@ -331,230 +124,6 @@ describe("FirstRunSetup", () => {
 
     expect(screen.getByTestId("first-run-repositories-loading")).toBeInTheDocument();
     expect(screen.queryByRole("option", { name: /octo\/stale-repo/ })).not.toBeInTheDocument();
-  });
-
-  it("moves the bound connection to another account through the picker", async () => {
-    const user = userEvent.setup();
-    const selectVcsConnection = vi.fn().mockResolvedValue(true);
-    bindMutate.mockImplementation((_vars: unknown, options: { onSuccess?: () => void }) => {
-      options.onSuccess?.();
-    });
-
-    renderSetup(boundConnectionModel(selectVcsConnection), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    await user.click(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("octo") }));
-
-    expect(bindMutate).toHaveBeenCalledWith(
-      { state: "csrf", installationId: "22" },
-      expect.objectContaining({ onSuccess: expect.any(Function) }),
-    );
-    await waitFor(() => expect(selectVcsConnection).toHaveBeenCalledWith("github-1"));
-    expect(await screen.findByTestId("first-run-choose")).toBeInTheDocument();
-  });
-
-  function bindablePageModel(selectVcsConnection: OnboardingPageModel["selectVcsConnection"]) {
-    const pendingInstance = {
-      metadata: { id: "int-new", integrationName: "github" },
-      status: {
-        state: "pending",
-        metadata: {
-          startedByUserID: "user-1",
-          state: "csrf",
-          githubApp: { slug: "superplane" },
-          pendingInstallations: [{ id: "11", accountLogin: "acme" }],
-        },
-      },
-    };
-    // The static test model shows the post-bind refetch already applied: the
-    // bound connection reports ready.
-    const readyInstance = {
-      metadata: { id: "int-new", name: "github-acme", integrationName: "github" },
-      status: { state: "ready", metadata: { owner: "acme" } },
-    };
-    return pageModel({
-      openSection: "vcs",
-      selectVcsConnection,
-      githubConnections: {
-        name: "github",
-        readyInstances: [readyInstance],
-        allInstances: [pendingInstance],
-      },
-    });
-  }
-
-  it("saves the bound GitHub connection for a workspace of an existing organization", async () => {
-    const user = userEvent.setup();
-    const selectVcsConnection = vi.fn().mockResolvedValue(true);
-    factory = { id: "factory-1", onboarding: {} };
-    bindMutate.mockImplementation((_vars: unknown, options: { onSuccess?: () => void }) => {
-      options.onSuccess?.();
-    });
-
-    renderSetup(bindablePageModel(selectVcsConnection), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    await user.click(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("acme") }));
-
-    await waitFor(() => expect(selectVcsConnection).toHaveBeenCalledWith("int-new"));
-    expect(await screen.findByTestId("first-run-choose")).toBeInTheDocument();
-  });
-
-  it("saves the bound GitHub connection for the initial organization", async () => {
-    const user = userEvent.setup();
-    const selectVcsConnection = vi.fn().mockResolvedValue(true);
-    factory = { id: "factory-1", onboarding: { initial: true } };
-    bindMutate.mockImplementation((_vars: unknown, options: { onSuccess?: () => void }) => {
-      options.onSuccess?.();
-    });
-
-    renderSetup(bindablePageModel(selectVcsConnection), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    await user.click(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("acme") }));
-
-    await waitFor(() => expect(selectVcsConnection).toHaveBeenCalledWith("int-new"));
-    expect(await screen.findByTestId("first-run-choose")).toBeInTheDocument();
-  });
-
-  // Regression: the repository screen opened before the bound connection was
-  // saved, so a fast repository pick stored the repository on the prior
-  // connection. The screen must stay on connect when the save fails.
-  it("keeps the connect screen when the bound connection does not save", async () => {
-    const user = userEvent.setup();
-    const selectVcsConnection = vi.fn().mockResolvedValue(false);
-    factory = { id: "factory-1", onboarding: {} };
-    bindMutate.mockImplementation((_vars: unknown, options: { onSuccess?: () => void }) => {
-      options.onSuccess?.();
-    });
-
-    renderSetup(bindablePageModel(selectVcsConnection), "/org-1/workspaces/PAY/setup?step=vcs");
-
-    await user.click(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("acme") }));
-
-    await waitFor(() => expect(selectVcsConnection).toHaveBeenCalledWith("int-new"));
-    expect(screen.getByTestId("first-run-connect")).toBeInTheDocument();
-    expect(screen.queryByTestId("first-run-choose")).not.toBeInTheDocument();
-  });
-
-  it("shows the GitHub account picker on the connect screen", () => {
-    renderSetup(
-      pageModel({
-        openSection: "vcs",
-        githubConnections: {
-          name: "github",
-          readyInstances: [],
-          allInstances: [
-            {
-              metadata: { id: "int-1", integrationName: "github" },
-              status: {
-                state: "pending",
-                metadata: {
-                  startedByUserID: "user-1",
-                  state: "csrf",
-                  githubApp: { slug: "superplane" },
-                  pendingInstallations: [
-                    { id: "11", accountLogin: "acme" },
-                    { id: "22", accountLogin: "octo" },
-                  ],
-                },
-              },
-            },
-          ],
-        },
-      }),
-      "/org-1/workspaces/PAY/setup?step=vcs",
-    );
-
-    expect(screen.getByRole("heading", { name: FIRST_RUN_COPY.connect.selectAccount })).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: FIRST_RUN_COPY.connect.useAccount("acme") })).toBeInTheDocument();
-    expect(screen.queryByTestId("first-run-connect-github")).not.toBeInTheDocument();
-  });
-
-  it("does not show another member's GitHub account picker", () => {
-    renderSetup(
-      pageModel({
-        openSection: "vcs",
-        githubConnections: {
-          name: "github",
-          readyInstances: [],
-          allInstances: [
-            {
-              metadata: { id: "int-1", integrationName: "github" },
-              status: {
-                state: "pending",
-                metadata: {
-                  startedByUserID: "some-other-user",
-                  state: "csrf",
-                  githubApp: { slug: "superplane" },
-                  pendingInstallations: [
-                    { id: "11", accountLogin: "acme" },
-                    { id: "22", accountLogin: "octo" },
-                  ],
-                },
-              },
-            },
-          ],
-        },
-      }),
-      "/org-1/workspaces/PAY/setup?step=vcs",
-    );
-
-    expect(screen.queryByTestId("first-run-github-account-picker")).not.toBeInTheDocument();
-    expect(screen.getByTestId("first-run-connect-github")).toBeInTheDocument();
-  });
-
-  it("opens Connect when GitHub returned an install request without a step", () => {
-    renderSetup(pageModel({ openSection: "vcs" }), "/org-1/workspaces/PAY/setup?githubSetup=request");
-
-    expect(screen.getByTestId("first-run-connect")).toBeInTheDocument();
-    expect(screen.getByTestId("first-run-github-install-requested")).toBeInTheDocument();
-    expect(screen.queryByTestId("first-run-welcome")).not.toBeInTheDocument();
-  });
-
-  it("shows a waiting chip when GitHub returned an install request", () => {
-    renderSetup(pageModel({ openSection: "vcs" }), "/org-1/workspaces/PAY/setup?step=vcs&githubSetup=request");
-
-    expect(screen.getByTestId("first-run-github-install-requested")).toHaveTextContent(
-      FIRST_RUN_COPY.connect.installRequested,
-    );
-    expect(screen.getByTestId("first-run-connect-github")).toBeInTheDocument();
-  });
-
-  it("shows a waiting account row when the pending connection has a request and no ready account", () => {
-    const pendingInstance = {
-      metadata: { id: "int-1", integrationName: "github" },
-      status: {
-        state: "pending",
-        metadata: {
-          startedByUserID: "user-1",
-          startedByGitHubLogin: "ada",
-          installRequested: true,
-          installRequestedAccount: "acme",
-          state: "csrf",
-          githubApp: { slug: "superplane" },
-        },
-      },
-    };
-
-    renderSetup(
-      pageModel({
-        openSection: "vcs",
-        githubConnections: { name: "github", allInstances: [pendingInstance], readyInstances: [] },
-      }),
-      "/org-1/workspaces/PAY/setup?step=vcs&githubSetup=request&githubOrg=acme",
-    );
-
-    expect(screen.getByTestId("first-run-github-account-picker")).toBeInTheDocument();
-    expect(screen.getByTestId("first-run-github-install-org")).toHaveTextContent("acme");
-    expect(screen.queryByTestId("first-run-connect-github")).not.toBeInTheDocument();
-  });
-
-  it("names the GitHub organization from the return query", () => {
-    renderSetup(
-      pageModel({ openSection: "vcs" }),
-      "/org-1/workspaces/PAY/setup?step=vcs&githubSetup=request&githubOrg=acme",
-    );
-
-    expect(screen.getByTestId("first-run-github-install-org")).toHaveTextContent("acme");
-    expect(screen.queryByText(FIRST_RUN_COPY.connect.installRequestedBody("acme"))).not.toBeInTheDocument();
   });
 
   it("counts the ticket screen as the last step when the agent screen is skipped", () => {
@@ -601,7 +170,7 @@ describe("FirstRunSetup", () => {
   });
 
   it("shows Log out and the organization switch when another workspace exists", () => {
-    factories = [factory, { id: "factory-2" }];
+    setupFixtures.factories = [setupFixtures.factory, { id: "factory-2" }];
 
     renderSetup(pageModel());
 
@@ -611,8 +180,8 @@ describe("FirstRunSetup", () => {
   });
 
   it("shows Log out and the organization switch when another organization exists", () => {
-    factories = [factory];
-    accountOrganizations = [
+    setupFixtures.factories = [setupFixtures.factory];
+    setupFixtures.accountOrganizations = [
       { id: "org-1", name: "Acme" },
       { id: "org-2", name: "Other Co" },
     ];
@@ -624,7 +193,7 @@ describe("FirstRunSetup", () => {
   });
 
   it("opens the current organization from the switch menu so the user can leave setup", async () => {
-    factories = [factory, { id: "factory-2" }];
+    setupFixtures.factories = [setupFixtures.factory, { id: "factory-2" }];
     const user = userEvent.setup();
 
     renderSetup(pageModel());
@@ -659,8 +228,8 @@ describe("FirstRunSetup", () => {
   });
 
   it("keeps Log out and hides the organization switch with a single org and single workspace", () => {
-    factories = [factory];
-    accountOrganizations = [{ id: "org-1", name: "Acme" }];
+    setupFixtures.factories = [setupFixtures.factory];
+    setupFixtures.accountOrganizations = [{ id: "org-1", name: "Acme" }];
 
     renderSetup(pageModel());
